refactor: UploadContainer를 페이지 전용 디렉토리로 이동
- 파일명 변경 (라우트 파일 아니므로 page.tsx는 부적절)
- features/upload/ 에서 app/label/upload/_components/로 이동
- Next.js App Router의 페이지 전용 컴포넌트 컨벤션 적용
